// client/src/models/IQuestion.ts
export interface IQuestion {
  id: string;
  question_text: string;
  question_picture: string;
  'arabic_translation of question': string;
  chapter_name_topic: string;
  duration: string;
  justification: string;
  hint: string;
  'arabic_translation of hint': string;
  'arabic_translation of justification': string;
  subject: string;
  skill_profile: string;
  question_type: string;
  sub_topic: string;
  content_type: string;
  difficulty_level: string;
  option_A: string;
  image_option_A: string;
  option_B: string;
  image_option_B: string;
  option_C: string;
  image_option_C: string;
  option_D: string;
  image_option_D: string;
  correct_options: string;
  input_type: string;
  justification_image: string;
  justification_type: string;
}